reward_platform/requirements.txt
Django==4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
celery
redis
gunicorn
drf_yasg
whitenoise==6.6.0
